LayoutTests/media/audio-mpeg-supported-expected.txt

Test that the audio element supports MPEG files.

EVENT(load)
EXPECTED (mediaElement.currentTime == '0') OK
EVENT(play)
EXPECTED (mediaElement.currentTime > '0') OK
END OF TEST
